Add import functionality with token persistence
- Add auth token persistence to database (survives restarts)
- Restore tokens from database on server startup
- Fix DuckDB UPSERT limitations (use DELETE+INSERT for indexed columns)
- Fix Strava API type mismatches (heartrate/watts as float64)
- Pass athlete ID to activity converter (required for foreign key)
- Use background context for async import (prevent early cancellation)
- Add import API endpoints (start, progress, cancel)
- Add import UI with progress display on settings page
- Create activities list and detail pages with TanStack Query
- Set up TanStack Router with route definitions
- Initialize shadcn/ui with Tailwind v4
- Create data layer with query key factories
- Add Zustand stores for settings and filters
- Update justfile to run both servers concurrently

Phase 4 (Activities Feature) is now complete.

+++ internal/api/handlers/activities.go
@@ -54,10 +54,10 @@ type ActivityResponse struct {
 	AverageSpeed         float64  `json:"average_speed"`
 	MaxSpeed             float64  `json:"max_speed"`
 	AverageHeartrate     *float64 `json:"average_heartrate,omitempty"`
-	MaxHeartrate         *int     `json:"max_heartrate,omitempty"`
+	MaxHeartrate         *float64 `json:"max_heartrate,omitempty"`
 	AverageWatts         *float64 `json:"average_watts,omitempty"`
-	MaxWatts             *int     `json:"max_watts,omitempty"`
-	WeightedAverageWatts *int     `json:"weighted_average_watts,omitempty"`
+	MaxWatts             *float64 `json:"max_watts,omitempty"`
+	WeightedAverageWatts *float64 `json:"weighted_average_watts,omitempty"`
 	Kilojoules           *float64 `json:"kilojoules,omitempty"`
 	AverageCadence       *float64 `json:"average_cadence,omitempty"`
 	Calories             *float64 `json:"calories,omitempty"`
